implement plan item instance level condition evaluation for remove wait_for_repetition case instance migration

+++ modules/flowable-cmmn-engine/src/main/java/org/flowable/cmmn/engine/impl/runtime/AbstractCmmnDynamicStateManager.java
@@ -465,15 +465,17 @@ public abstract class AbstractCmmnDynamicStateManager {
                 for (PlanItemInstance planItemInstance : planItemInstances) {
                     if (planItemInstance.getState().equalsIgnoreCase(PlanItemInstanceState.WAITING_FOR_REPETITION)) {
                         PlanItemInstanceEntity planItemInstanceEntity = (PlanItemInstanceEntity) planItemInstance;
-                        
-                        Date currentTime = cmmnEngineConfiguration.getClock().getCurrentTime();
-                        CmmnHistoryManager cmmnHistoryManager = cmmnEngineConfiguration.getCmmnHistoryManager();
-                        planItemInstanceEntity.setState(PlanItemInstanceState.TERMINATED);
-                        planItemInstanceEntity.setEndedTime(currentTime);
-                        planItemInstanceEntity.setTerminatedTime(currentTime);
-                        cmmnHistoryManager.recordPlanItemInstanceTerminated(planItemInstanceEntity);
-                        
-                        planItemInstanceEntityManager.delete(planItemInstanceEntity);
+
+                        if (evaluateCondition(planItemInstanceEntity, planItemDefinitionMapping)) {
+                            Date currentTime = cmmnEngineConfiguration.getClock().getCurrentTime();
+                            CmmnHistoryManager cmmnHistoryManager = cmmnEngineConfiguration.getCmmnHistoryManager();
+                            planItemInstanceEntity.setState(PlanItemInstanceState.TERMINATED);
+                            planItemInstanceEntity.setEndedTime(currentTime);
+                            planItemInstanceEntity.setTerminatedTime(currentTime);
+                            cmmnHistoryManager.recordPlanItemInstanceTerminated(planItemInstanceEntity);
+
+                            planItemInstanceEntityManager.delete(planItemInstanceEntity);
+                        }
                     }
                 }
             }
